adjust tenant

+++ database/migrations/2025_10_28_000100_rename_tenant_column_on_user_tenants_table.php
@@ -13,13 +13,16 @@ return new class extends Migration
     {
         if (Schema::hasTable('user_tenants')) {
             // drop the column tenant
-            Schema::table('user_tenants', function (Blueprint $table) {
-                $table->dropColumn('tenant');
-            });
-
-            Schema::table('user_tenants', function (Blueprint $table) {
-                $table->foreignId('tenant_id')->constrained('tenants')->cascadeOnDelete()->cascadeOnUpdate();
-            });
+            if (Schema::hasColumn('user_tenants', 'tenant')) {
+                Schema::table('user_tenants', function (Blueprint $table) {
+                    $table->dropColumn('tenant');
+                });
+            }
+            if (!Schema::hasColumn('user_tenants', 'tenant_id')) {
+                Schema::table('user_tenants', function (Blueprint $table) {
+                    $table->foreignId('tenant_id')->constrained('tenants')->cascadeOnDelete()->cascadeOnUpdate();
+                });
+            }
         }
     }
